Replace .tif and .jpg files with .tga
This means I can get rid of the dependencies on libtiff, libjpeg, libpng
(we weren't using any .png files anyway) and libz (which libpng depends
on). Win!

Also removed a lot of image loading code that wasn't being used (and
won't be, since we're going for .tga for all our images). I've left the
.bmp loading code in for now, since it has no external dependencies, but
I should probably remove that too.