Capitalise image file names and invert tgas
TGA textures were upside-down, because the row order in those files is
the opposite of what we were getting from the .tifs. I've put in a quick
fix which scales the texture matrix by -1 along the y axis to make them
show up correctly, but a better fix would be to store the rows in the
correct order as they're read in.

Note that not all of the image file names have been capitalised in this
check-in, because OS X's stupid filesystem pretends to be case sensitive
but really isn't. This will cause runtime errors in linux at the moment,
but should still work OK on OS X.